refactor: extract column analysis logic into column_analysis.py
